requirements.txt
python-telegram-bot
flask
requests
aiohttp
gunicorn
python-dotenv 